Revert "feat: move additional helpers"
