+++ package.json
@@ -39,8 +39,5 @@
     "react-dom": "19.0.0",
     "typescript": "~5.7.2",
     "vitest": "^2.1.6"
-  },
-  "dependencies": {
-    "openai": "^4.91.1"
   }
 }
