+++ src/types/conversationHistories.ts
@@ -1,0 +1,9 @@
+export type ConversationHistory = {
+  id: string;
+  model: string;
+  title: string;
+  lastSaved: number;
+  tokensRemaining: number;
+};
+
+export type ConversationHistories = Record<string, ConversationHistory>;
